Improve the digraph module.
Fix a mistake in the comment on an exported predicate.
Improve the wording in other comments.

Use explicit module qualification in bunch of places, because without them
references to predicate names that occur in many standard library modules
(that digraph.h imports) are ambiguous to human readers, though not to
the compiler.

Replace some functions with state-variable-friendly predicates.