Introduce and use iom_item_and_specs.
Some of the errors that we can discover during parsing of items
are recoverable, in the sense that after generating an error message
for the problem, we can still construct an item that is correct
in all other respects, and which we can use to build a HLDS on which
we can perform semantic analysis without the results being misleading.

We have traditionally done this by threading !Specs through both
parse_item_or_marker and the predicates that call it. However, this
makes it look as though adding to !Specs is a *normal* activity
for parse_item_or_marker, in the sense that regardless of which kind
of item or marker a particular call to parse_item_or_marker returns,
it may also have added something to !Specs as well. This is false.
Most values in the item_or_marker type represent markers, and
if parse_item_or_marker returns *any* of them, it *cannot* also
add anything to !Specs.

Change the interface of parse_item_or_marker so that it does not take
a !Specs pair of arguments. Instead, extend the item_or_marker type
with a new function symbol, iom_item_and_specs, which includes
the error_specs of any recovered errors alongside the item during whose
processing they were discovered.

This changed interface makes it possible to refactor the code handling
module starts in a way that makes it clear that getting any error_specs
at all is NOT part of the usual path through that code (since .m files that
do *not* begin with a ":- module" declaration are of course very rare).

compiler/parse_types.m:
    Add iom_item_and_specs to the item_or_marker type.

compiler/parse_type_defn.m:
    When parsing type definitions, return any error_specs generated
    for recoverable errors using iom_item_and_specs. (Type definition items
    are the only kinds of items at present for which some errors
    are recoverable.)

compiler/parse_item.m:
    Do not pass !Specs through the code parsing items and markers.

compiler/parse_module.m:
    Implement the refactor mentioned above.

compiler/parse_class.m:
    Conform to the changes above.