Simplify the printing of class declarations.
Previously, the three parts of a class declaration line
(the initial indent, the flags, and the enum or struct declaration)
were printed in three separate places. Make mlds_output_class_decl
responsible for printing them all, as well as any #line directive
we may want to print before the declaration.

Make mlds_output_class_defn responsible for printing the whole
of a class definition. Previously, it did part of that job,
then handed off the rest to mlds_output_class *in the middle of the line*
containing the class declaration.

Give a dead predicate a descriptive name.