Avoid a C# error with bitwise complement.
C# will promote the bitwise complement of byte or ushort into an int and then
sometimes complains about possible information loss when we cast it back to the
original type.  (Specifically, it complains when the operand is a literal value
whose complement is a negative int.)  This is causing compilation of the
standard library to fail with -O5 --intermod-opt.  The fix is to wrap the
bitwise complement in an unchecked context.

compiler/mlds_to_cs_data.m:
    As above.