Specialise file copying in the compiler.
Specialise the Mercury implementation of file copying in the compiler by
avoiding the use of io.binary_input_stream_foldl_io/5. This allows us to avoid
a higher-order call each time a byte is written and it allows us to use the
unboxed version of the predicate that reads bytes. The other reason for this
change is that we are planning to deprecate (and eventually remove)
io.binary_input_stream_foldl_io/5.

compiler/module_cmds.m:
    As above.

    Add a comment about improving the Mercury implementation of file copying.

    Fix spelling in a spot.

library/io.m:
    Refer to the Boolean constant yes/0 and no/0 in C code as MR_YES
    and MR_NO respectively, rather than MR_TRUE and MR_FALSE.