Fix the failure of debugger/interpreter in debug grades.
In debug grades, io.close_input is traced, so an mdb single-step command
stops there; in non-debug grades, io.close_input is not traced, so
single-step does not stop there. This was the cause of the difference
in output.

Make this cause irrelevant by changing the relevant mdb command in the input
from single-step to finish of the parent, which goes to the same event
in both debug and non-debug grades.