Add an XXX about a probable bug.
Inline a predicate at its four call sites, since the resulting code
is just as easy to read as the original.

Factor four occurrences of a call into one call, and then inline it,
since it is now the only call to the callee.

Use more consistent variable and predicate names.