Overhaul of the Syntax chapter of the reference manual
There are three main aims for this change:

1. Don't rely so much on users having knowledge of Prolog (ideally we
wouldn't require it at all, but this change probably doesn't fully achieve
that). Present the information more linearly, i.e., without requiring
so many forward references.

2. Make important information less likely to be overlooked. As it stands,
we give a lot of semantics in the Syntax chapter where users wouldn't
expect to find it (they're more likely to just skim over such a chapter).

3. Remove some inconsistencies in our descriptions and use of terminology,
and try to remove ambiguities.

Changes to terminology
----------------------

These are aimed at being more consistent, as well as dividing things up
into concepts that we need to refer to later on.

- A term is now a core term or a special term. The latter are for list
syntax, etc. Avoid the use of "is parsed as", the significance of which
is not clear, and instead define an equivalence relation via a term
normalization procedure.

- Define "functor" and "principal functor". Don't use "functor" to talk
about terms: the functor is defined as the name/arity pair. Also give
some synonyms we use for functor.

- Core terms are divided into variables, constants, and functor terms.
Functor terms are names on their own or names with arguments
(i.e., "functor terms" are those terms that have meaningful functors).

- "Higher-order terms" are now called "apply terms", since they aren't
necessarily higher-order themselves, and because that is a semantic
concept anyway.

- Data-terms are now referred to as expressions.

- Refer to "field access" rather than "record syntax". It's not made clear
that these are referring to the same feature, so it's better to stick to
one term (at least until full definitions are given).

Changes to document structure
-----------------------------

The Syntax chapter as a whole has been split into two chapters,
Syntax and Clauses. The new structure is as follows:

Syntax

- Additional sections are added for comments and line number directives.

- The Tokens section is split into token types (so that, e.g., variables
can be looked up quickly).

- "Operators" covers everything about operator syntax, not just builtin
operators.

- "Terms" covers the material previously in that section. It adds a summary
in the form of grammar rules (but makes it clear all the information is
also in the prose), and a term normalization procedure that is _not_
written in pseudo-Mercury. The grammar is there for people who are likely
to skip the prose but less likely to skip a grammar, and the prose is there
because that's the best way to describe things.

- "Items" covers the material previously in "Items", "Declarations",
"Facts", and "Rules".

Clauses

- Added an overview of Mercury semantics.

- "Goals" covers the same material as before, but I have rearranged the
order of the goals so as to avoid forward references, and to put more
commonly used goals earlier.

- "Expressions" covers the material previously in "Data-terms" and its
subsections. The expressions are (mostly) collected into one table, but
field access expressions are kept separate as they require a bit more
explanation. I've used an anchor on this separate bit, however, so maybe
it should be in a section that follows Expressions (i.e., Expressions
should say "see next section for details" or whatever).

- "State variables", "Variable scoping", "Implicit quantifiaction", and
"Elimination of double negation" are kept largely as is.

- "Definite Clause Grammars" is added which collects all the information
relating to DCGs.

Other changes
-------------

- Names now include sequences of graphic chars and a single semicolon,
which is what our parser actually allows.

- Use hyphens in meta-variables rather than underscore, per the texinfo
guidelines.

- Explain that operators are a syntactic concept and you need to import
a module to get, e.g., arithmetic operations. Point out the user defined
operator in the prose rather than in a footnote. And fix the formatting
of that table!

- Try to use a more consistent writing style.