Inline throw_on_error and related preds ...
... so that compiler optimizations can separate the usual success path,
which does nothing, from the error path, which constructs and throws
an exception.

Modify throw_on_error itself to manually inline is_error, so that
we can (a) eliminate redundant testing for success/failure, and
(b) move all work that is unneeded on the success path out of the
success path. (Marking is_error to be inlined is not sufficient
to do this with our current optimization setup.)