Optimise dependency file index maps.
In make_info we maintain a mapping between dependency_file and
dependency_file_index. Replace the type dependency_file used in those
maps with a new type, in which modules are referred to by module_index
instead of by module_name.

The reason for the change is that make.dependencies.of_3 is called
frequently to convert a module_index and module_target_type to a
dependency_file_index. With this change, it does not need to look up
the module_name corresponding to a module_index, but uses the
module_index as given. Also, using module_index instead of module_name
as part of a hash table key should be faster as (1) hashing is faster,
and (2) comparing keys within the same hash table bucket is faster.

This change improves the run time of a do-nothing build of Prince
on my machine from 1.8 s to 1.7 s.

compiler/make.dependencies.m:
    Add the new type.

    Don't need to call module_index_to_name in of_3.

    Describe the `of' function.

compiler/make.make_info.m:
    Replace dependency_file in the dependency file <-> index maps
    with the new type.

compiler/make.util.m:
    Add hash predicate for new type.

compiler/make.deps_set.m:
compiler/make.top_level.m:
    Conform to changes.