Cache timestamps by target file.
Previously, to look up the timestamp for a target_file we would first
compute a file name for that target_file, then look up the timestamp by
file name. This occurs frequently, so even though the computed file name
was cached, eliminating one of the two steps can be worthwhile.

This change therefore introduces a cache of timestamps indexed by
target_file, in addition to the existing cache of timestamps indexed
by file name.

Cache invalidation can be complicated because when a file is updated,
it is not straightforward to which target_file(s) may be affected.
Invalidating the entire cache is viable as it happens relatively
infrequently, and the cache will be quickly repopulated anyway.

This change improves the run time of a do-nothing build of Prince
on my machine from 1.7 s to 1.6 s.

compiler/make.make_info.m:
compiler/make.top_level.m:
    Add a field to make_info to hold a cache of timestamps indexed by
    target_file.

    Delete the mki_search_file_name_cache field.

compiler/make.util.m:
    Cache timestamps by target_file in get_target_timestamp.

    Avoid computing file names until required.

    Don't cache computed file names in get_file_name.
    get_file_name is now called much less frequently, so caching the
    result should not be necessary (but can be reintroduced if some
    other code path requires it).

    Clear out timestamps for target_files when a file is deleted.

compiler/make.module_target.m:
    Delete old timestamp for a target_file when the target is made.

compiler/make.program_target.m:
    Clear out timestamps for target_files when Java class files are
    made.