Ensure trans_opt_deps rules can be read back.
When writing out a trans_opt_deps rule in a .d file,
always write each dependency on a separate line. If the rule is
written "horizontally" because there is only one dependency, e.g.

    foo.trans_opt_date : bar.trans_opt

then maybe_read_d_file_for_trans_opt_deps will not be able to parse the
list, and the .trans_opt file will not be used.

compiler/mmakefiles.m:
    As above.