Fix Github issue #116.
Building programs with --intermodule-optimization fails when using a library
containing submodules that was *not* installed with --intermodule-optimization
enabled. This is because the compiler is trying to find the .int0 files for the
library, but these are only installed when the library is built with
--intermodule-optimization enabled. The fix is to always install any .int0
files for a library.

compiler/make.program_target.m:
compiler/write_deps_file.m:
    Always install the .int0 files for libraries.