Fix generation of the debugger documentation on Windows.
We generate the debugger documentation files, e.g. the command list, by
processing the output of the info tool and running through a series of sh and
awk scripts. This is breaking builds for MSYS based systems on Windows because
the port of awk there does not handle non-ASCII characters and info will
sometimes output those. Using the C locale on affected systems causes the
output of info to be ASCII.

doc/generate_mdb_doc:
    Force the use of the C locale on MSYS / MinGW systems in order to avoid
    problems with awk.

    Shut up some warnings form shellcheck.