Do not include the fullarch in version and usage messages.
The first line of usage and version messages have long been concluded with "on
<ARCH>" component, where <ARCH> is the value of FULLARCH baked into the
standard library. Omit this component, it doesn't convey any useful information
and we need to make some space on the first line for the optional package
version that will be added in an upcoming change.

compiler/handle_options.m:
profiler/mercury_profile.m:
slice/mcov.m:
     As above.