Mark operations on deps_sets as such.
We have traditionally defined deps_sets to be sparse_bitsets, but left
operations on them without module qualification to allow this to be changed
quickly and simply, e.g. to plain sets for debugging. Unfortunately,
this makes code that operates on both deps_sets and other kinds of sets
hard to read. Fix this.

compiler/make.deps_set.m:
    Add a set of predicates and functions that operate on deps_sets.
    These all forward the work directly to the corresponding predicates
    and functions from sparse_bitset.m, or from whichever other set module
    we want to use to represent deps_sets.

    Replace unqualified calls that operate on deps_sets with calls to the
    new operations.

    Make predicate and variable names more meaningful.

compiler/make.dependencies.m:
    Replace unqualified calls that operate on deps_sets with calls to the
    new operations.

    Replace unqualified calls that operate on plain sets with qualified calls.

    Eliminate an intermediate data structure, and give the affected predicate
    a meaningful name.