Add a new option --show-movability.
This new takes pred (or func) names as arguments. It is an accumulating option,
i.e. it can be used to specify several such names. The effect is to ask
the compiler to generate a report on the effect of moving the specified
predicates and/or functions to a new module, by

- listing the other predicates and/or functions that would need to be moved
  with them, and
- warning about any unwanted coupling that such a move would introduce
  between the old and the new modules.

compiler/options.m:
    Add the new option.

compiler/hlds_call_tree.m:
    Add code to implement the new option to this new module, since
    it already has code to do the bulk of the work, which consists of
    computing the local pred_ids accessible from each local pred_info.

    Factor out the code computing this info into a predicate of its own,
    to make it possible to invoke both the old functionality of this module
    and the new without doing that work twice.

compiler/mercury_compile_front_end.m:
    Invoke the new code in hlds_call_tree.m if the new option is given,
    and update the code for invoking the old functionality in hlds_call_tree.m.