Fix update of timestamp cache by target file.
We introduced a cache of timestamps by target file in commit 4d4ef3e3a,
but when a target is made, only the timestamp for the main TargetFile
was invalidated, not the timestamps of other files that may also have
been touched, such as files for any nested submodules.

Fixes Mantis bug #569.

compiler/make.module_target.m:
    Delete the timestamps of all TouchedTargetFiles in
    record_made_target_given_maybe_touched_files,
    not only the main TargetFile.