Reset target file timestamp cache before installing library grade.
When installing library grades with mmc --make, we change the grade and
enable the use of grade subdirs (if not already enabled) in the globals.
We must also clear the target file timestamp cache, as the information
it contains is only valid for a given grade and grade-subdir setting.

compiler/make.program_target.m:
    As above.

compiler/make.timestamp.m:
    Add an optional trace goal to check that a timestamp retrieved
    from the cache matches the actual timestamp of the file on disk.

    Rename get_target_timestamp_2 to get_target_timestamp_uncached.